api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "REST API service"
authors = ["NeoTheBestDeveloper <neothebestmain@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
drf-spectacular = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
select = ["ALL"]
ignore = [
    "D1", # No docstrings.
    "ANN101", # Do not annotate Self.
    "RUF001", # False positive. Tell about latin C and cyrilc C

    "FBT", # I want to use boolean as a function argument.
    "A003", # I want to use 'id' as name for object attribute.
    "D205", # I don't want to write \n before summary line in docstring.
    "RUF002", # I don't care that several russian letters look like some english letters.
]
fix = true
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$" # For unusable variables.

[tool.mypy]
python_version = "3.12"
strict = true
show_error_codes = true
ignore_missing_imports = true # Work incorrectly and useless
disallow_untyped_decorators = false # Because of FastAPI and aiogramm decorators invoke this error
disallow_subclassing_any = false

[tool.pyright]
include = ["src"]
strict = ["src"]
reportMissingImports = false